refactor: 1차 피드백 개선
Common 파일 -> Fruit, Juice, JuiceMakingError 으로 나눔
errorMessage를 JuiceMakingError 내부에 존재하도록 변경
errorMessage 출력함수 생성
FruitStore에 name 프로퍼티 삭제
