+++ JuiceMaker/JuiceMaker/Model/Fruit.swift
@@ -1,0 +1,12 @@
+//
+//  Fruit.swift
+//  JuiceMaker
+//
+//  Created by Vayne, 우롱차 on 2022/02/24.
+//
+
+import Foundation
+
+enum Fruit {
+    case strawberry, banana, pineapple, kiwi, mango
+}
